+++ src/main/java/com/real/backend/domain/auth/controller/AuthController.java
@@ -1,4 +1,4 @@
-package com.real.backend.domain.oauth.controller;
+package com.real.backend.domain.auth.controller;
 
 import java.io.IOException;
 
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.real.backend.domain.oauth.kakao.KakaoUtil;
-import com.real.backend.domain.oauth.service.AuthService;
+import com.real.backend.domain.auth.kakao.KakaoUtil;
+import com.real.backend.domain.auth.service.AuthService;
 import com.real.backend.response.StatusResponse;
 
 import jakarta.servlet.http.HttpServletRequest;
